Migrate tests to correct file
